requirements.txt
discord.py==1.3.2
PyNaCl==1.3.0